bitmap: fix problems in previous commit
Commit ee3d3893 missed the fact that (unsigned char)<<(int)
is truncated to int, and therefore failed for any bitmap data
longer than four bytes.

Also, I failed to run 'make syntax-check' on my commit 4bba6579;
for whatever odd reason, ffs lives in a different header than ffsl.

* src/util/bitmap.c (virBitmapNewData): Use correct shift type.
(includes): Glibc (and therefore gnulib) decided ffs is in
<strings.h>, but ffsl is in <string.h>.
* tests/virbitmaptest.c (test5): Test it.

+++ src/util/bitmap.c
@@ -27,7 +27,6 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
-#include <strings.h>
 #include <stdlib.h>
 #include <sys/types.h>
 
@@ -434,7 +433,7 @@ virBitmapPtr virBitmapNewData(void *data, int len)
             j = 0;
             p++;
         }
-        *p |= bytes[i] << (j * CHAR_BIT);
+        *p |= (unsigned long) bytes[i] << (j * CHAR_BIT);
     }
 
     return bitmap;
